rename: diaryList 폴더 구조를 diary로 변경 후 list와 content폴더로 나눔
