+++ src/features/diary/apis/index.ts
@@ -4,7 +4,7 @@ import {
   File,
   MetaData,
   PatchFileArgs,
-} from '@/features/diaryList/apis/interfaces.ts';
+} from '@/features/diary/apis/interfaces.ts';
 import driveApi from '@/features/common/apis/driveInstance.ts';
 import driveUploadApi from '@/features/common/apis/driveUploadInstance.ts';
