+++ src/features/diary/apis/mutations/usePatchFile.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
-import { patchFile } from '@/features/diaryList/apis/index.ts';
-import { PatchFileArgs } from '@/features/diaryList/apis/interfaces.ts';
+import { patchFile } from '@/features/diary/apis/index.ts';
+import { PatchFileArgs } from '@/features/diary/apis/interfaces.ts';
 import { DIARY_KEY } from '@/constants/queryKey.ts';
 
 const usePatchFile = () =>
