+++ src/features/diary/apis/queries/useGetFileList.ts
@@ -1,6 +1,6 @@
-import { getFileList } from '@/features/diaryList/apis/index.ts';
+import { getFileList } from '@/features/diary/apis/index.ts';
 import { useQuery } from '@tanstack/react-query';
-import { DriveFile } from '@/features/diaryList/apis/interfaces.ts';
+import { DriveFile } from '@/features/diary/apis/interfaces.ts';
 import { DIARY_KEY } from '@/constants/queryKey.ts';
 
 const useGetFileList = (accessToken?: string) =>
