+++ src/features/diary/apis/queries/useGetFileMetaData.ts
@@ -1,6 +1,6 @@
 import { DIARY_KEY } from '@/constants/queryKey.ts';
-import { getFileMetaData } from '@/features/diaryList/apis/index.ts';
-import { MetaData } from '@/features/diaryList/apis/interfaces.ts';
+import { getFileMetaData } from '@/features/diary/apis/index.ts';
+import { MetaData } from '@/features/diary/apis/interfaces.ts';
 import { useQuery } from '@tanstack/react-query';
 
 const useGetFileMetaData = (id: string, accessToken?: string) =>
