+++ src/features/diary/components/content/DiaryEditor/plugins/InitialPlugin.tsx
