+++ src/features/diary/components/list/DiaryCard/DiaryCard.tsx
@@ -1,10 +1,10 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import { File } from '@/features/diaryList/apis/interfaces.ts';
-import * as style from '@/features/diaryList/components/DiaryCard/DiaryCard.css.ts';
+import { File } from '@/features/diary/apis/interfaces.ts';
+import * as style from '@/features/diary/components/list/DiaryCard/DiaryCard.css.ts';
 import earth from '@/static/diary/earth2.svg';
 import calendar from '@/static/diary/calendar.svg';
-import DiaryCardDropdown from '@/features/diaryList/components/DiaryCard/DiaryCardDropdown/DiaryCardDropdown.tsx';
+import DiaryCardDropdown from '@/features/diary/components/list/DiaryCardDropdown/DiaryCardDropdown.tsx';
 import { MouseEvent } from 'react';
 
 export default function DiaryCard({ file }: { file: File }) {
