+++ src/features/diary/components/list/DiaryCardDropdown/DiaryCardDropdown.tsx
@@ -1,8 +1,8 @@
 import menu from '@/static/diary/menu.svg';
 import Image from 'next/image';
-import useDeleteFile from '@/features/diaryList/apis/mutations/useDeleteFile.ts';
+import useDeleteFile from '@/features/diary/apis/mutations/useDeleteFile.ts';
 import Dropdown from '@/features/common/components/dropdown/Dropdown.tsx';
-import * as style from '@/features/diaryList/components/DiaryCard/DiaryCardDropdown/DiaryCardDropdown.css.ts';
+import * as style from '@/features/diary/components/list/DiaryCardDropdown/DiaryCardDropdown.css.ts';
 
 export default function DiaryCardDropdown({ id }: { id: string }) {
   const deleteMutation = useDeleteFile();
