+++ src/pages/_app.tsx
@@ -15,7 +15,7 @@ import bootApp from '@/core/config/boostApp.ts';
 import PageLayout from '@/features/common/components/PageLayout.tsx';
 
 import '@/styles/globals.css.ts';
-import '@/features/diaryList/components/DiaryEditor/lexical.css';
+import '@/features/diary/components/content/DiaryEditor/lexical.css';
 import { isSSR } from '@/core/utils/objectUtils.ts';
 
 type NextPageWithLayout = NextPage & {
