+++ src/pages/diary/[id].tsx
@@ -1,9 +1,9 @@
 import { GetServerSideProps } from 'next';
-import { MetaData } from '@/features/diaryList/apis/interfaces.ts';
+import { MetaData } from '@/features/diary/apis/interfaces.ts';
 import { QueryClient, dehydrate } from '@tanstack/react-query';
-import { getFile, getFileMetaData } from '@/features/diaryList/apis/index.ts';
+import { getFile, getFileMetaData } from '@/features/diary/apis/index.ts';
 import { DIARY_KEY } from '@/constants/queryKey.ts';
-export { default } from '@/features/diaryList/pages/content/DiaryContentPage.tsx';
+export { default } from '@/features/diary/pages/content/DiaryContentPage.tsx';
 
 export const getServerSideProps: GetServerSideProps = async context => {
   const queryClient = new QueryClient();
