+++ src/pages/diary/index.tsx
@@ -1,9 +1,9 @@
 import { DIARY_KEY } from '@/constants/queryKey.ts';
-import { getFileList } from '@/features/diaryList/apis/index.ts';
-import { DriveFile } from '@/features/diaryList/apis/interfaces.ts';
+import { getFileList } from '@/features/diary/apis/index.ts';
+import { DriveFile } from '@/features/diary/apis/interfaces.ts';
 import { QueryClient, dehydrate } from '@tanstack/react-query';
 import { GetServerSideProps } from 'next';
-export { default } from '@/features/diaryList/pages/root/DiaryPage.tsx';
+export { default } from '@/features/diary/pages/root/DiaryPage.tsx';
 
 export const getServerSideProps: GetServerSideProps = async context => {
   const queryClient = new QueryClient();
